+++ Packages/BrandonUtils/Standalone/Clerical/ClericalErrors.cs
@@ -4,6 +4,8 @@ using BrandonUtils.Standalone.Strings;
 
 namespace BrandonUtils.Standalone.Clerical {
     public static class ClericalErrors {
+        #region Already Exists
+
         private static IOException _itemAlreadyExistsException(object item) {
             return new IOException($"The {item.GetType().Prettify()} {item.Prettify()} already exists!");
         }
@@ -16,8 +18,18 @@ namespace BrandonUtils.Standalone.Clerical {
             return _itemAlreadyExistsException(item);
         }
 
+        #endregion
+
+        #region Does Not Exist
+
         private static IOException _itemDoesNotExistException(object item) {
-            return new IOException($"The {item.GetType().Prettify()} {item.Prettify()} does not exist!");
+            var message = $"The {item.GetType().Prettify()} {item.Prettify()} does not exist!";
+
+            return item switch {
+                FileInfo fi      => new FileNotFoundException(message, fi.FullName),
+                DirectoryInfo di => new DirectoryNotFoundException(message),
+                _                => new IOException(message)
+            };
         }
 
         public static IOException ItemDoesNotExistException(FileSystemInfo item) {
@@ -27,5 +39,7 @@ namespace BrandonUtils.Standalone.Clerical {
         public static IOException ItemDoesNotExistException(IHasFileSystemInfo item) {
             return _itemAlreadyExistsException(item);
         }
+
+        #endregion
     }
 }