+++ Packages/BrandonUtils/Standalone/Clerical/Saving/SaveFileName.cs
@@ -34,20 +34,31 @@ namespace BrandonUtils.Standalone.Clerical.Saving {
         private static readonly  RegexGroup NicknameGroup       = new RegexGroup(nameof(NicknameGroup),  @".+");
         private static readonly  RegexGroup TimeStampGroup      = new RegexGroup(nameof(TimeStampGroup), @"\d+");
         internal static readonly Regex      BaseFileNamePattern = new Regex($@"^{NicknameGroup}_{TimeStampGroup}$");
+        internal static readonly Regex      FileNamePattern     = new Regex($@"^{NicknameGroup}_{TimeStampGroup}{BPath.ExtensionGroup}$");
+
+        internal static string GetFileSearchPattern(string nickname, string fullExtension) {
+            return $"{nickname}_*{fullExtension}";
+        }
+
+        internal string GetFileSearchPattern() => GetFileSearchPattern(Nickname, FullExtension);
 
         public static SaveFileName Parse(string fileName) {
-            var parsed   = new SaveFileName();
-            var baseName = BPath.GetFileNameWithoutExtensions(fileName);
-            var match    = BaseFileNamePattern.Match(baseName);
+            return _parseInternal(fileName, FileNamePattern);
+        }
+
+        private static SaveFileName _parseInternal(string fileName, Regex expectedPattern) {
+            var match = expectedPattern.Match(fileName);
+
             if (match.Success) {
-                parsed.Nickname = match.Groups[NicknameGroup.GroupName].Value;
-
-                var tsString = match.Groups[TimeStampGroup.GroupName].Value;
-                parsed.TimeStamp = new DateTime(long.Parse(tsString));
+                return new SaveFileName() {
+                    Nickname      = match.Groups[NicknameGroup.GroupName].Value,
+                    TimeStamp     = new DateTime(long.Parse(match.Groups[TimeStampGroup.GroupName].Value)),
+                    FullExtension = match.Groups[NicknameGroup.GroupName].Value,
+                };
             }
-
-            parsed.FullExtension = BPath.GetFullExtension(fileName);
-            return parsed;
+            else {
+                throw new FormatException($"Unable to convert the string \"{fileName.ToString(default)}\" to a {nameof(SaveFileName)} because it didn't match the {nameof(Regex)} pattern /{expectedPattern}/");
+            }
         }
 
         public static SaveFileName Parse(FileInfo fileInfo) {
@@ -66,6 +77,10 @@ namespace BrandonUtils.Standalone.Clerical.Saving {
             }
 
             return nickname;
+        }
+
+        public override string ToString() {
+            return Rendered;
         }
     }
 }