+++ Packages/BrandonUtils/Standalone/Clerical/Saving/SaveFolder.cs
@@ -1,4 +1,6 @@
-﻿using System.IO;
+﻿using System.Collections.Generic;
+using System.IO;
+using System.Linq;
 
 using JetBrains.Annotations;
 
@@ -20,6 +22,18 @@ namespace BrandonUtils.Standalone.Clerical.Saving {
             GameName
         );
 
-        public override DirectoryInfo DirectoryInfo => new DirectoryInfo(FolderPath);
+        public override DirectoryInfo Directory => new DirectoryInfo(FolderPath);
+
+        [NotNull]
+        public IEnumerable<SaveFile<T>> EnumerateSaveFiles<T>(string searchPattern) where T : ISaveData {
+            if (Exists) {
+                return EnumerateFiles(searchPattern, SearchOption.TopDirectoryOnly)
+                       .Select(it => new SaveFile<T>(this, it))
+                       .OrderByDescending(it => it.TimeStamp);
+            }
+            else {
+                return Enumerable.Empty<SaveFile<T>>();
+            }
+        }
     }
 }